Add feature to control backups on metered connections
restic-windows-backup runs backups irrespective of the host being on a metered connection or unrestricted. That can lead to situations in which the backup takes place when the user is on a metered connection and would rather not do it at a time, either because they would incurr costs or use up their data allowance.

The default value of $BackupOnMeteredNetwork is set to $false following a conservative approach.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>